#30 - Remove old forms, controller and converters
